Double Base Converter (Base 2 - 62)
Works for all bases except there is one problem, if the number is greater than 17 digits any digits after it get cut off because a double is a 16bit signature so it only carries the scientific number decimal up to the 16th place and then cuts it off and applies the power. This means any digits at the end get cut off and the end digits of the conversion end up being incorrect while the beginning digits still retain their correct values.